requirements.txt
telethon==1.36.0
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.0.0
colorlog>=6.7.0
httpx[socks]>=0.27.0
google-genai>=0.1.0
deepl>=1.17.0
openai>=1.0.0
anthropic>=0.34.0
supabase>=2.0.0
langchain-core>=0.3.0
langgraph>=0.1.8
pytest-asyncio>=0.24.0
